requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
redis[hiredis]==5.0.1
aiohttp==3.9.1

# AI/ML Libraries
torch==2.1.2
transformers==4.37.0
faster-whisper==0.10.0
TTS==0.22.0
fasttext-wheel==0.9.2

# Audio Processing
numpy==1.26.3
soundfile==0.12.1
librosa==0.10.1

# Monitoring
prometheus-client==0.19.0
